some changes

+++ src/main/java/sorting/ShellSort.java
@@ -2,7 +2,7 @@ package sorting;
 
 public class ShellSort {
 
-  public void sort(int[] arrayToSort) {
+  public void sort(double[] arrayToSort) {
     validateParams(arrayToSort);
     int n = arrayToSort.length;
 
@@ -11,7 +11,7 @@ public class ShellSort {
 
     for (int gap = n / 2; gap > 0; gap /= 2) {
       for (int i = gap; i < n; i++) {
-        int key = arrayToSort[i];
+        double key = arrayToSort[i];
         int j = i;
         while (j >= gap && arrayToSort[j - gap] > key) {
           arrayToSort[j] = arrayToSort[j - gap];
@@ -22,7 +22,7 @@ public class ShellSort {
     }
   }
 
-  private void validateParams(int array_param[]) {
+  private void validateParams(double array_param[]) {
 
   }
 }
